requirements.txt
django
whitenoise
pillow
django-jazzmin
django-import-export
langchain
langchain-google-genai 
langchain-community
chromadb
pypdf
schedule
